requirements.txt
streamlit
google-generativeai
Pillow
watchdog